requirements.txt
# DSPy and GEPA
dspy==3.1.2
gepa>=0.0.24

# LLM APIs and Unified Interface
litellm>=1.80.10
openai>=2.14.0

# Configuration and Environment
python-dotenv>=1.2.1
PyYAML>=6.0.3

# Data Analysis and Visualization
matplotlib
pandas
numpy